Noise reduction using signal to noise ratio (SNR)
Now a days we are facing lots of lots of disturbance with background noise and all the real
measurements are disturbed by noise. This includes electronic noise, but can also include
An external event that affects the measured phenomenon like – wind, vibrations the
Gravitational attraction of the moon, variations of temperature, variations of humidity etc.,
depending on what is measured and of the sensitivity of the device. It is often possible to
reduce the noise by controlling the environment.
Internal electronic noise of measurement systems can be reduced through the use of low noise
Amplifiers .When the characteristics of the noise are known and are different from the signal it
is possible to use a filter to reduce the noise. A low pass filter (LPF) is an audio signal processor
that removes unwanted noise frequencies from a signal above a determined cut off frequency.
All real measurements are disturbed by noise. This includes electronic noise, but can also include
external events that affect the measured phenomenon — wind, vibrations, the gravitational attraction
of the moon, variations of temperature, variations of humidity, etc., depending on what is measured
and of the sensitivity of the device. It is often possible to reduce the noise by controlling the
environment.
Internal electronic noise of measurement systems can be reduced through the use of low-noise
Amplifiers. When the characteristics of the noise are known and are different from the signal,
it is possible to use a filter to reduce the noise.